Implement something similar to raganwalds transitionsTo style machine
This is nice in the sense that we can clearly see what the transitions are, and in which states
they're valid. It's a bit meh because we sacrifice a type safety (in the sense that all the
possible transitions are available on the base state, so there's no way for the compiler to
know whether a transition is legal - it will just blow up at runtime).