Enforce transitions which return the current type
Using a recursive type definition we can declare on the base class that the deposit method returns the "current state" where the current state is whatever is given in the generic type parameter.
This helps avoid the bug where depositing balance in an Open account could accidentally transition the state Held.